+++ partials/sidebar.html.php
@@ -9,7 +9,7 @@
             <i class="fas fa-tachometer-alt mr-3"></i>
             Dashboard
         </a>
-        <a href="query/party.php" class="flex items-center text-white opacity-75 hover:opacity-100 py-4 pl-6 nav-item">
+        <a href="parties.html.php" class="flex items-center text-white opacity-75 hover:opacity-100 py-4 pl-6 nav-item">
             <i class="fas fa-align-left mr-3"></i>
             Party
         </a>
@@ -30,6 +30,10 @@
         </a>
 
     </nav>
+
+    <div>
+        <strong>ADMIN</strong>
+    </div>
 
 </aside>
