+++ query/register.controller.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 require_once('validations.php');
@@ -33,5 +34,4 @@ function signUpUser(
 
 
     $statement->execute();
-
 }
